backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
anthropic>=0.7.0
openai>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
pandas>=2.1.0
pillow>=10.1.0
numpy>=1.26.0
supabase>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
